SQL Workshop

+++ financialCalculatorProject/src/main/java/org/example/financialCalculator3.java
@@ -12,10 +12,12 @@ public class financialCalculator3 {
         double rate = scanner.nextDouble();
         System.out.println("Number of years since initial deposit?");
         int years = scanner.nextInt();
-        int n = (int) (12 * years);
+        int n = (12 * years);
         double r = rate/12/100;
         double pv  = payment * ( Math.pow(1 - (1 + r),(-n)) / r);
         System.out.printf("Present value of annuity is " + " %.2f",pv);
+
+
